backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2023.7.22
cffi
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.58.0
django-celery-beat==2.6.0
django-cors-headers==4.3.0
django-smtp-ssl==1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.7
multidict==6.0.5
oauthlib==3.2.2
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.6
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.4.post2
pillow==10.4.0
weasyprint==62.3
sslcommerz-lib==1.0
pdf417gen==0.8.0
pandas
openpyxl

